CSCQVAIN:242: REF: Adding Ajv code into seperate file
